requirements.txt
discord.py==2.5.2
aiosqlite==0.21.0
python-dotenv>=1.0,<2.0